requirements.txt
Flask==2.3.2
pyjwt==2.8.0
cryptography==41.0.3
python-dotenv==1.0.0
redis==4.5.5
bcrypt==4.0.1
flask-cors==4.0.0
numpy==1.26.4
pandas==2.2.2
requests==2.31.0
openai==0.27.8